requirements.txt
Django==4.0.1
python-dotenv==0.20.0
django-environ==0.9.0
psycopg2-binary==2.9.3
djangorestframework==3.13.1
social-auth-app-django==5.0.0
requests==2.9.1
drf-yasg==1.21.3
urllib3==1.26.10
django-filter==22.1

